// src/libs/api/v2/model/rates.ts
/**
 * Fixer.io
 * Foreign Exchange Rates and Currency Conversion API (http://fixer.io/)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface Rates {
  base?: string;
  date?: string;
  rates?: { [key: string]: number };
}